organize report mod; it's just nicer
